+++ packages/form_companion_presenter/test/async_validation_indicator_test.dart
@@ -4,7 +4,6 @@
 
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -68,112 +67,6 @@ class Presenter with CompanionPresenterMixin, FormCompanionMixin {
 
   @override
   FutureOr<void> doSubmit() => _doSubmitCalled();
-}
-
-class DummyBuildContext extends BuildContext {
-  DummyBuildContext();
-
-  @override
-  bool get debugDoingBuild => throw UnimplementedError();
-
-  @override
-  InheritedWidget dependOnInheritedElement(
-    InheritedElement ancestor, {
-    Object? aspect,
-  }) {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? dependOnInheritedWidgetOfExactType<T extends InheritedWidget>({
-    Object? aspect,
-  }) {
-    // Required for getLocale() test.
-    return null;
-  }
-
-  @override
-  DiagnosticsNode describeElement(
-    String name, {
-    DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty,
-  }) {
-    throw UnimplementedError();
-  }
-
-  @override
-  List<DiagnosticsNode> describeMissingAncestor({
-    required Type expectedAncestorType,
-  }) {
-    throw UnimplementedError();
-  }
-
-  @override
-  DiagnosticsNode describeOwnershipChain(String name) {
-    throw UnimplementedError();
-  }
-
-  @override
-  DiagnosticsNode describeWidget(
-    String name, {
-    DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty,
-  }) {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorRenderObjectOfType<T extends RenderObject>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorStateOfType<T extends State<StatefulWidget>>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorWidgetOfExactType<T extends Widget>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  RenderObject? findRenderObject() {
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findRootAncestorStateOfType<T extends State<StatefulWidget>>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  InheritedElement?
-      getElementForInheritedWidgetOfExactType<T extends InheritedWidget>() {
-    throw UnimplementedError();
-  }
-
-  @override
-  BuildOwner? get owner => throw UnimplementedError();
-
-  @override
-  Size? get size => throw UnimplementedError();
-
-  @override
-  void visitAncestorElements(bool Function(Element element) visitor) {
-    throw UnimplementedError();
-  }
-
-  @override
-  void visitChildElements(ElementVisitor visitor) {
-    throw UnimplementedError();
-  }
-
-  @override
-  Widget get widget => throw UnimplementedError();
-
-  @override
-  void dispatchNotification(Notification notification) {
-    throw UnimplementedError();
-  }
 }
 
 void main() {
